+++ apps/desktop/src/app/session/hooks/use-session-list-actions.test.tsx
@@ -98,7 +98,8 @@ vi.mock('@/store/gateway', async importOriginal => ({
 // the whole projects store (gateway / fs / git) into this hook's test.
 const removed = vi.hoisted(() => ({ ids: new Set<string>() }))
 
-vi.mock('@/store/projects', () => ({
+vi.mock('@/store/session-removal', async importActual => ({
+  ...(await importActual<Record<string, unknown>>()),
   $removedSessionIds: { get: () => removed.ids }
 }))
